refactor: 리팩토링에 따른 불러오기 변경
